backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.25.2
sqlalchemy==2.0.23
alembic==1.13.1
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
websockets==12.0
redis==5.0.1
structlog==23.2.0
pydantic-settings==2.1.0